@@ -1,10 +1,10 @@
 name: flutter_reactive_ble_example
 description: Demonstrates how to use the flutter_reactive_ble plugin.
-version: 2.7.3
+version: 3.1.0
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
@@ -12,10 +12,15 @@ dependencies:
     sdk: flutter
   flutter_reactive_ble:
     path: ../
-  provider: ^4.0.4
+  provider: ^5.0.0
+  functional_data: ^1.0.0
+  intl: ^0.17.0
+
 
 dev_dependencies:
-  dependency_validator:
+  build_runner: ^1.10.11
+  dependency_validator: ^2.0.1
+  functional_data_generator: ^1.0.0-nullsafety.1 
   flutter_test:
     sdk: flutter
 
