@@ -1,29 +1,28 @@
 name: flutter_reactive_ble
 description: Reactive Bluetooth Low Energy (BLE) plugin that can communicate with multiple devices
-version: 2.7.3
+version: 3.1.0
 homepage: https://github.com/PhilipsHue/flutter_reactive_ble
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
-  collection: ^1.14.12
+  collection: ^1.15.0
   flutter:
     sdk: flutter
-  functional_data: ^0.3.0
-  meta: ^1.1.8
-  pedantic: ^1.9.0
-  protobuf: ^1.0.1
+  functional_data: ^1.0.0
+  meta: ^1.3.0
+  pedantic: ^1.11.0
+  protobuf: ^2.0.0
 
 dev_dependencies:
-  build_runner: ^1.9.0
-  dependency_validator: ^1.4.2
+  build_runner: ^1.12.1
   flutter_test:
     sdk: flutter
-  functional_data_generator: ^0.3.0
-  mockito: ^4.1.3
-
+  functional_data_generator: ^1.0.0
+  mockito: ^5.0.2
+  
 flutter:
   plugin:
     platforms:
