@@ -2,7 +2,7 @@
 //  Generated code. Do not modify.
 //  source: bledata.proto
 //
-// @dart = 2.3
+// @dart = 2.12
 // ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields
 
 import 'dart:core' as $core;
@@ -32,7 +32,23 @@ class ScanForDevicesRequest extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   ScanForDevicesRequest._() : super();
-  factory ScanForDevicesRequest() => create();
+  factory ScanForDevicesRequest({
+    $core.Iterable<Uuid>? serviceUuids,
+    $core.int? scanMode,
+    $core.bool? requireLocationServicesEnabled,
+  }) {
+    final _result = create();
+    if (serviceUuids != null) {
+      _result.serviceUuids.addAll(serviceUuids);
+    }
+    if (scanMode != null) {
+      _result.scanMode = scanMode;
+    }
+    if (requireLocationServicesEnabled != null) {
+      _result.requireLocationServicesEnabled = requireLocationServicesEnabled;
+    }
+    return _result;
+  }
   factory ScanForDevicesRequest.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -49,8 +65,8 @@ class ScanForDevicesRequest extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   ScanForDevicesRequest copyWith(
           void Function(ScanForDevicesRequest) updates) =>
-      super.copyWith((message) => updates(
-          message as ScanForDevicesRequest)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as ScanForDevicesRequest))
+          as ScanForDevicesRequest; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static ScanForDevicesRequest create() => ScanForDevicesRequest._();
@@ -60,7 +76,7 @@ class ScanForDevicesRequest extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static ScanForDevicesRequest getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<ScanForDevicesRequest>(create);
-  static ScanForDevicesRequest _defaultInstance;
+  static ScanForDevicesRequest? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.List<Uuid> get serviceUuids => $_getList(0);
@@ -103,8 +119,7 @@ class DeviceScanInfo extends $pb.GeneratedMessage {
         const $core.bool.fromEnvironment('protobuf.omit_field_names')
             ? ''
             : 'name')
-    ..aOM<GenericFailure>(
-        3, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'failure',
+    ..aOM<GenericFailure>(3, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'failure',
         subBuilder: GenericFailure.create)
     ..pc<ServiceDataEntry>(
         4, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'serviceData', $pb.PbFieldType.PM,
@@ -116,14 +131,46 @@ class DeviceScanInfo extends $pb.GeneratedMessage {
             : 'rssi',
         $pb.PbFieldType.O3)
     ..a<$core.List<$core.int>>(
-        6,
-        const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'manufacturerData',
-        $pb.PbFieldType.OY,
+        6, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'manufacturerData', $pb.PbFieldType.OY,
         protoName: 'manufacturerData')
+    ..pc<Uuid>(7, const $core.bool.fromEnvironment('protobuf.omit_field_names') ? '' : 'serviceUuids', $pb.PbFieldType.PM,
+        protoName: 'serviceUuids', subBuilder: Uuid.create)
     ..hasRequiredFields = false;
 
   DeviceScanInfo._() : super();
-  factory DeviceScanInfo() => create();
+  factory DeviceScanInfo({
+    $core.String? id,
+    $core.String? name,
+    GenericFailure? failure,
+    $core.Iterable<ServiceDataEntry>? serviceData,
+    $core.int? rssi,
+    $core.List<$core.int>? manufacturerData,
+    $core.Iterable<Uuid>? serviceUuids,
+  }) {
+    final _result = create();
+    if (id != null) {
+      _result.id = id;
+    }
+    if (name != null) {
+      _result.name = name;
+    }
+    if (failure != null) {
+      _result.failure = failure;
+    }
+    if (serviceData != null) {
+      _result.serviceData.addAll(serviceData);
+    }
+    if (rssi != null) {
+      _result.rssi = rssi;
+    }
+    if (manufacturerData != null) {
+      _result.manufacturerData = manufacturerData;
+    }
+    if (serviceUuids != null) {
+      _result.serviceUuids.addAll(serviceUuids);
+    }
+    return _result;
+  }
   factory DeviceScanInfo.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -138,8 +185,8 @@ class DeviceScanInfo extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   DeviceScanInfo copyWith(void Function(DeviceScanInfo) updates) =>
-      super.copyWith((message) =>
-          updates(message as DeviceScanInfo)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as DeviceScanInfo))
+          as DeviceScanInfo; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static DeviceScanInfo create() => DeviceScanInfo._();
@@ -149,7 +196,7 @@ class DeviceScanInfo extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static DeviceScanInfo getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<DeviceScanInfo>(create);
-  static DeviceScanInfo _defaultInstance;
+  static DeviceScanInfo? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.String get id => $_getSZ(0);
@@ -215,6 +262,9 @@ class DeviceScanInfo extends $pb.GeneratedMessage {
   $core.bool hasManufacturerData() => $_has(5);
   @$pb.TagNumber(6)
   void clearManufacturerData() => clearField(6);
+
+  @$pb.TagNumber(7)
+  $core.List<Uuid> get serviceUuids => $_getList(6);
 }
 
 class ConnectToDeviceRequest extends $pb.GeneratedMessage {
@@ -241,7 +291,24 @@ class ConnectToDeviceRequest extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   ConnectToDeviceRequest._() : super();
-  factory ConnectToDeviceRequest() => create();
+  factory ConnectToDeviceRequest({
+    $core.String? deviceId,
+    ServicesWithCharacteristics? servicesWithCharacteristicsToDiscover,
+    $core.int? timeoutInMs,
+  }) {
+    final _result = create();
+    if (deviceId != null) {
+      _result.deviceId = deviceId;
+    }
+    if (servicesWithCharacteristicsToDiscover != null) {
+      _result.servicesWithCharacteristicsToDiscover =
+          servicesWithCharacteristicsToDiscover;
+    }
+    if (timeoutInMs != null) {
+      _result.timeoutInMs = timeoutInMs;
+    }
+    return _result;
+  }
   factory ConnectToDeviceRequest.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -258,8 +325,8 @@ class ConnectToDeviceRequest extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   ConnectToDeviceRequest copyWith(
           void Function(ConnectToDeviceRequest) updates) =>
-      super.copyWith((message) => updates(
-          message as ConnectToDeviceRequest)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as ConnectToDeviceRequest))
+          as ConnectToDeviceRequest; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static ConnectToDeviceRequest create() => ConnectToDeviceRequest._();
@@ -269,7 +336,7 @@ class ConnectToDeviceRequest extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static ConnectToDeviceRequest getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<ConnectToDeviceRequest>(create);
-  static ConnectToDeviceRequest _defaultInstance;
+  static ConnectToDeviceRequest? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.String get deviceId => $_getSZ(0);
@@ -339,7 +406,23 @@ class DeviceInfo extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   DeviceInfo._() : super();
-  factory DeviceInfo() => create();
+  factory DeviceInfo({
+    $core.String? id,
+    $core.int? connectionState,
+    GenericFailure? failure,
+  }) {
+    final _result = create();
+    if (id != null) {
+      _result.id = id;
+    }
+    if (connectionState != null) {
+      _result.connectionState = connectionState;
+    }
+    if (failure != null) {
+      _result.failure = failure;
+    }
+    return _result;
+  }
   factory DeviceInfo.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -354,8 +437,8 @@ class DeviceInfo extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   DeviceInfo copyWith(void Function(DeviceInfo) updates) =>
-      super.copyWith((message) =>
-          updates(message as DeviceInfo)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as DeviceInfo))
+          as DeviceInfo; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static DeviceInfo create() => DeviceInfo._();
@@ -364,7 +447,7 @@ class DeviceInfo extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static DeviceInfo getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<DeviceInfo>(create);
-  static DeviceInfo _defaultInstance;
+  static DeviceInfo? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.String get id => $_getSZ(0);
@@ -420,7 +503,15 @@ class DisconnectFromDeviceRequest extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   DisconnectFromDeviceRequest._() : super();
-  factory DisconnectFromDeviceRequest() => create();
+  factory DisconnectFromDeviceRequest({
+    $core.String? deviceId,
+  }) {
+    final _result = create();
+    if (deviceId != null) {
+      _result.deviceId = deviceId;
+    }
+    return _result;
+  }
   factory DisconnectFromDeviceRequest.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -437,8 +528,9 @@ class DisconnectFromDeviceRequest extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   DisconnectFromDeviceRequest copyWith(
           void Function(DisconnectFromDeviceRequest) updates) =>
-      super.copyWith((message) => updates(message
-          as DisconnectFromDeviceRequest)); // ignore: deprecated_member_use
+      super.copyWith(
+              (message) => updates(message as DisconnectFromDeviceRequest))
+          as DisconnectFromDeviceRequest; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static DisconnectFromDeviceRequest create() =>
@@ -449,7 +541,7 @@ class DisconnectFromDeviceRequest extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static DisconnectFromDeviceRequest getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<DisconnectFromDeviceRequest>(create);
-  static DisconnectFromDeviceRequest _defaultInstance;
+  static DisconnectFromDeviceRequest? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.String get deviceId => $_getSZ(0);
@@ -479,7 +571,15 @@ class ClearGattCacheRequest extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   ClearGattCacheRequest._() : super();
-  factory ClearGattCacheRequest() => create();
+  factory ClearGattCacheRequest({
+    $core.String? deviceId,
+  }) {
+    final _result = create();
+    if (deviceId != null) {
+      _result.deviceId = deviceId;
+    }
+    return _result;
+  }
   factory ClearGattCacheRequest.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -496,8 +596,8 @@ class ClearGattCacheRequest extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   ClearGattCacheRequest copyWith(
           void Function(ClearGattCacheRequest) updates) =>
-      super.copyWith((message) => updates(
-          message as ClearGattCacheRequest)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as ClearGattCacheRequest))
+          as ClearGattCacheRequest; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static ClearGattCacheRequest create() => ClearGattCacheRequest._();
@@ -507,7 +607,7 @@ class ClearGattCacheRequest extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static ClearGattCacheRequest getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<ClearGattCacheRequest>(create);
-  static ClearGattCacheRequest _defaultInstance;
+  static ClearGattCacheRequest? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.String get deviceId => $_getSZ(0);
@@ -537,7 +637,15 @@ class ClearGattCacheInfo extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   ClearGattCacheInfo._() : super();
-  factory ClearGattCacheInfo() => create();
+  factory ClearGattCacheInfo({
+    GenericFailure? failure,
+  }) {
+    final _result = create();
+    if (failure != null) {
+      _result.failure = failure;
+    }
+    return _result;
+  }
   factory ClearGattCacheInfo.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -552,8 +660,8 @@ class ClearGattCacheInfo extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   ClearGattCacheInfo copyWith(void Function(ClearGattCacheInfo) updates) =>
-      super.copyWith((message) => updates(
-          message as ClearGattCacheInfo)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as ClearGattCacheInfo))
+          as ClearGattCacheInfo; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static ClearGattCacheInfo create() => ClearGattCacheInfo._();
@@ -563,7 +671,7 @@ class ClearGattCacheInfo extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static ClearGattCacheInfo getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<ClearGattCacheInfo>(create);
-  static ClearGattCacheInfo _defaultInstance;
+  static ClearGattCacheInfo? _defaultInstance;
 
   @$pb.TagNumber(1)
   GenericFailure get failure => $_getN(0);
@@ -595,7 +703,15 @@ class NotifyCharacteristicRequest extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   NotifyCharacteristicRequest._() : super();
-  factory NotifyCharacteristicRequest() => create();
+  factory NotifyCharacteristicRequest({
+    CharacteristicAddress? characteristic,
+  }) {
+    final _result = create();
+    if (characteristic != null) {
+      _result.characteristic = characteristic;
+    }
+    return _result;
+  }
   factory NotifyCharacteristicRequest.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -612,8 +728,9 @@ class NotifyCharacteristicRequest extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   NotifyCharacteristicRequest copyWith(
           void Function(NotifyCharacteristicRequest) updates) =>
-      super.copyWith((message) => updates(message
-          as NotifyCharacteristicRequest)); // ignore: deprecated_member_use
+      super.copyWith(
+              (message) => updates(message as NotifyCharacteristicRequest))
+          as NotifyCharacteristicRequest; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static NotifyCharacteristicRequest create() =>
@@ -624,7 +741,7 @@ class NotifyCharacteristicRequest extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static NotifyCharacteristicRequest getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<NotifyCharacteristicRequest>(create);
-  static NotifyCharacteristicRequest _defaultInstance;
+  static NotifyCharacteristicRequest? _defaultInstance;
 
   @$pb.TagNumber(1)
   CharacteristicAddress get characteristic => $_getN(0);
@@ -656,7 +773,15 @@ class NotifyNoMoreCharacteristicRequest extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   NotifyNoMoreCharacteristicRequest._() : super();
-  factory NotifyNoMoreCharacteristicRequest() => create();
+  factory NotifyNoMoreCharacteristicRequest({
+    CharacteristicAddress? characteristic,
+  }) {
+    final _result = create();
+    if (characteristic != null) {
+      _result.characteristic = characteristic;
+    }
+    return _result;
+  }
   factory NotifyNoMoreCharacteristicRequest.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -673,8 +798,9 @@ class NotifyNoMoreCharacteristicRequest extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   NotifyNoMoreCharacteristicRequest copyWith(
           void Function(NotifyNoMoreCharacteristicRequest) updates) =>
-      super.copyWith((message) => updates(message
-          as NotifyNoMoreCharacteristicRequest)); // ignore: deprecated_member_use
+      super.copyWith((message) =>
+              updates(message as NotifyNoMoreCharacteristicRequest))
+          as NotifyNoMoreCharacteristicRequest; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static NotifyNoMoreCharacteristicRequest create() =>
@@ -686,7 +812,7 @@ class NotifyNoMoreCharacteristicRequest extends $pb.GeneratedMessage {
   static NotifyNoMoreCharacteristicRequest getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<NotifyNoMoreCharacteristicRequest>(
           create);
-  static NotifyNoMoreCharacteristicRequest _defaultInstance;
+  static NotifyNoMoreCharacteristicRequest? _defaultInstance;
 
   @$pb.TagNumber(1)
   CharacteristicAddress get characteristic => $_getN(0);
@@ -718,7 +844,15 @@ class ReadCharacteristicRequest extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   ReadCharacteristicRequest._() : super();
-  factory ReadCharacteristicRequest() => create();
+  factory ReadCharacteristicRequest({
+    CharacteristicAddress? characteristic,
+  }) {
+    final _result = create();
+    if (characteristic != null) {
+      _result.characteristic = characteristic;
+    }
+    return _result;
+  }
   factory ReadCharacteristicRequest.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -735,8 +869,8 @@ class ReadCharacteristicRequest extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   ReadCharacteristicRequest copyWith(
           void Function(ReadCharacteristicRequest) updates) =>
-      super.copyWith((message) => updates(message
-          as ReadCharacteristicRequest)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as ReadCharacteristicRequest))
+          as ReadCharacteristicRequest; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static ReadCharacteristicRequest create() => ReadCharacteristicRequest._();
@@ -746,7 +880,7 @@ class ReadCharacteristicRequest extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static ReadCharacteristicRequest getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<ReadCharacteristicRequest>(create);
-  static ReadCharacteristicRequest _defaultInstance;
+  static ReadCharacteristicRequest? _defaultInstance;
 
   @$pb.TagNumber(1)
   CharacteristicAddress get characteristic => $_getN(0);
@@ -787,7 +921,23 @@ class CharacteristicValueInfo extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   CharacteristicValueInfo._() : super();
-  factory CharacteristicValueInfo() => create();
+  factory CharacteristicValueInfo({
+    CharacteristicAddress? characteristic,
+    $core.List<$core.int>? value,
+    GenericFailure? failure,
+  }) {
+    final _result = create();
+    if (characteristic != null) {
+      _result.characteristic = characteristic;
+    }
+    if (value != null) {
+      _result.value = value;
+    }
+    if (failure != null) {
+      _result.failure = failure;
+    }
+    return _result;
+  }
   factory CharacteristicValueInfo.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -804,8 +954,8 @@ class CharacteristicValueInfo extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   CharacteristicValueInfo copyWith(
           void Function(CharacteristicValueInfo) updates) =>
-      super.copyWith((message) => updates(
-          message as CharacteristicValueInfo)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as CharacteristicValueInfo))
+          as CharacteristicValueInfo; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static CharacteristicValueInfo create() => CharacteristicValueInfo._();
@@ -815,7 +965,7 @@ class CharacteristicValueInfo extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static CharacteristicValueInfo getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<CharacteristicValueInfo>(create);
-  static CharacteristicValueInfo _defaultInstance;
+  static CharacteristicValueInfo? _defaultInstance;
 
   @$pb.TagNumber(1)
   CharacteristicAddress get characteristic => $_getN(0);
@@ -879,7 +1029,19 @@ class WriteCharacteristicRequest extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   WriteCharacteristicRequest._() : super();
-  factory WriteCharacteristicRequest() => create();
+  factory WriteCharacteristicRequest({
+    CharacteristicAddress? characteristic,
+    $core.List<$core.int>? value,
+  }) {
+    final _result = create();
+    if (characteristic != null) {
+      _result.characteristic = characteristic;
+    }
+    if (value != null) {
+      _result.value = value;
+    }
+    return _result;
+  }
   factory WriteCharacteristicRequest.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -896,8 +1058,9 @@ class WriteCharacteristicRequest extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   WriteCharacteristicRequest copyWith(
           void Function(WriteCharacteristicRequest) updates) =>
-      super.copyWith((message) => updates(message
-          as WriteCharacteristicRequest)); // ignore: deprecated_member_use
+      super.copyWith(
+              (message) => updates(message as WriteCharacteristicRequest))
+          as WriteCharacteristicRequest; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static WriteCharacteristicRequest create() => WriteCharacteristicRequest._();
@@ -907,7 +1070,7 @@ class WriteCharacteristicRequest extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static WriteCharacteristicRequest getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<WriteCharacteristicRequest>(create);
-  static WriteCharacteristicRequest _defaultInstance;
+  static WriteCharacteristicRequest? _defaultInstance;
 
   @$pb.TagNumber(1)
   CharacteristicAddress get characteristic => $_getN(0);
@@ -957,7 +1120,19 @@ class WriteCharacteristicInfo extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   WriteCharacteristicInfo._() : super();
-  factory WriteCharacteristicInfo() => create();
+  factory WriteCharacteristicInfo({
+    CharacteristicAddress? characteristic,
+    GenericFailure? failure,
+  }) {
+    final _result = create();
+    if (characteristic != null) {
+      _result.characteristic = characteristic;
+    }
+    if (failure != null) {
+      _result.failure = failure;
+    }
+    return _result;
+  }
   factory WriteCharacteristicInfo.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -974,8 +1149,8 @@ class WriteCharacteristicInfo extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   WriteCharacteristicInfo copyWith(
           void Function(WriteCharacteristicInfo) updates) =>
-      super.copyWith((message) => updates(
-          message as WriteCharacteristicInfo)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as WriteCharacteristicInfo))
+          as WriteCharacteristicInfo; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static WriteCharacteristicInfo create() => WriteCharacteristicInfo._();
@@ -985,7 +1160,7 @@ class WriteCharacteristicInfo extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static WriteCharacteristicInfo getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<WriteCharacteristicInfo>(create);
-  static WriteCharacteristicInfo _defaultInstance;
+  static WriteCharacteristicInfo? _defaultInstance;
 
   @$pb.TagNumber(1)
   CharacteristicAddress get characteristic => $_getN(0);
@@ -1038,7 +1213,19 @@ class NegotiateMtuRequest extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   NegotiateMtuRequest._() : super();
-  factory NegotiateMtuRequest() => create();
+  factory NegotiateMtuRequest({
+    $core.String? deviceId,
+    $core.int? mtuSize,
+  }) {
+    final _result = create();
+    if (deviceId != null) {
+      _result.deviceId = deviceId;
+    }
+    if (mtuSize != null) {
+      _result.mtuSize = mtuSize;
+    }
+    return _result;
+  }
   factory NegotiateMtuRequest.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1053,8 +1240,8 @@ class NegotiateMtuRequest extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   NegotiateMtuRequest copyWith(void Function(NegotiateMtuRequest) updates) =>
-      super.copyWith((message) => updates(
-          message as NegotiateMtuRequest)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as NegotiateMtuRequest))
+          as NegotiateMtuRequest; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static NegotiateMtuRequest create() => NegotiateMtuRequest._();
@@ -1064,7 +1251,7 @@ class NegotiateMtuRequest extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static NegotiateMtuRequest getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<NegotiateMtuRequest>(create);
-  static NegotiateMtuRequest _defaultInstance;
+  static NegotiateMtuRequest? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.String get deviceId => $_getSZ(0);
@@ -1113,7 +1300,23 @@ class NegotiateMtuInfo extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   NegotiateMtuInfo._() : super();
-  factory NegotiateMtuInfo() => create();
+  factory NegotiateMtuInfo({
+    $core.String? deviceId,
+    $core.int? mtuSize,
+    GenericFailure? failure,
+  }) {
+    final _result = create();
+    if (deviceId != null) {
+      _result.deviceId = deviceId;
+    }
+    if (mtuSize != null) {
+      _result.mtuSize = mtuSize;
+    }
+    if (failure != null) {
+      _result.failure = failure;
+    }
+    return _result;
+  }
   factory NegotiateMtuInfo.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1128,8 +1331,8 @@ class NegotiateMtuInfo extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   NegotiateMtuInfo copyWith(void Function(NegotiateMtuInfo) updates) =>
-      super.copyWith((message) => updates(
-          message as NegotiateMtuInfo)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as NegotiateMtuInfo))
+          as NegotiateMtuInfo; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static NegotiateMtuInfo create() => NegotiateMtuInfo._();
@@ -1139,7 +1342,7 @@ class NegotiateMtuInfo extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static NegotiateMtuInfo getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<NegotiateMtuInfo>(create);
-  static NegotiateMtuInfo _defaultInstance;
+  static NegotiateMtuInfo? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.String get deviceId => $_getSZ(0);
@@ -1195,7 +1398,15 @@ class BleStatusInfo extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   BleStatusInfo._() : super();
-  factory BleStatusInfo() => create();
+  factory BleStatusInfo({
+    $core.int? status,
+  }) {
+    final _result = create();
+    if (status != null) {
+      _result.status = status;
+    }
+    return _result;
+  }
   factory BleStatusInfo.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1210,8 +1421,8 @@ class BleStatusInfo extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   BleStatusInfo copyWith(void Function(BleStatusInfo) updates) =>
-      super.copyWith((message) =>
-          updates(message as BleStatusInfo)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as BleStatusInfo))
+          as BleStatusInfo; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static BleStatusInfo create() => BleStatusInfo._();
@@ -1221,7 +1432,7 @@ class BleStatusInfo extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static BleStatusInfo getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<BleStatusInfo>(create);
-  static BleStatusInfo _defaultInstance;
+  static BleStatusInfo? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.int get status => $_getIZ(0);
@@ -1259,7 +1470,19 @@ class ChangeConnectionPriorityRequest extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   ChangeConnectionPriorityRequest._() : super();
-  factory ChangeConnectionPriorityRequest() => create();
+  factory ChangeConnectionPriorityRequest({
+    $core.String? deviceId,
+    $core.int? priority,
+  }) {
+    final _result = create();
+    if (deviceId != null) {
+      _result.deviceId = deviceId;
+    }
+    if (priority != null) {
+      _result.priority = priority;
+    }
+    return _result;
+  }
   factory ChangeConnectionPriorityRequest.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1276,8 +1499,9 @@ class ChangeConnectionPriorityRequest extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   ChangeConnectionPriorityRequest copyWith(
           void Function(ChangeConnectionPriorityRequest) updates) =>
-      super.copyWith((message) => updates(message
-          as ChangeConnectionPriorityRequest)); // ignore: deprecated_member_use
+      super.copyWith(
+              (message) => updates(message as ChangeConnectionPriorityRequest))
+          as ChangeConnectionPriorityRequest; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static ChangeConnectionPriorityRequest create() =>
@@ -1289,7 +1513,7 @@ class ChangeConnectionPriorityRequest extends $pb.GeneratedMessage {
   static ChangeConnectionPriorityRequest getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<ChangeConnectionPriorityRequest>(
           create);
-  static ChangeConnectionPriorityRequest _defaultInstance;
+  static ChangeConnectionPriorityRequest? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.String get deviceId => $_getSZ(0);
@@ -1337,7 +1561,19 @@ class ChangeConnectionPriorityInfo extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   ChangeConnectionPriorityInfo._() : super();
-  factory ChangeConnectionPriorityInfo() => create();
+  factory ChangeConnectionPriorityInfo({
+    $core.String? deviceId,
+    GenericFailure? failure,
+  }) {
+    final _result = create();
+    if (deviceId != null) {
+      _result.deviceId = deviceId;
+    }
+    if (failure != null) {
+      _result.failure = failure;
+    }
+    return _result;
+  }
   factory ChangeConnectionPriorityInfo.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1354,8 +1590,9 @@ class ChangeConnectionPriorityInfo extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   ChangeConnectionPriorityInfo copyWith(
           void Function(ChangeConnectionPriorityInfo) updates) =>
-      super.copyWith((message) => updates(message
-          as ChangeConnectionPriorityInfo)); // ignore: deprecated_member_use
+      super.copyWith(
+              (message) => updates(message as ChangeConnectionPriorityInfo))
+          as ChangeConnectionPriorityInfo; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static ChangeConnectionPriorityInfo create() =>
@@ -1366,7 +1603,7 @@ class ChangeConnectionPriorityInfo extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static ChangeConnectionPriorityInfo getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<ChangeConnectionPriorityInfo>(create);
-  static ChangeConnectionPriorityInfo _defaultInstance;
+  static ChangeConnectionPriorityInfo? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.String get deviceId => $_getSZ(0);
@@ -1421,7 +1658,23 @@ class CharacteristicAddress extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   CharacteristicAddress._() : super();
-  factory CharacteristicAddress() => create();
+  factory CharacteristicAddress({
+    $core.String? deviceId,
+    Uuid? serviceUuid,
+    Uuid? characteristicUuid,
+  }) {
+    final _result = create();
+    if (deviceId != null) {
+      _result.deviceId = deviceId;
+    }
+    if (serviceUuid != null) {
+      _result.serviceUuid = serviceUuid;
+    }
+    if (characteristicUuid != null) {
+      _result.characteristicUuid = characteristicUuid;
+    }
+    return _result;
+  }
   factory CharacteristicAddress.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1438,8 +1691,8 @@ class CharacteristicAddress extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   CharacteristicAddress copyWith(
           void Function(CharacteristicAddress) updates) =>
-      super.copyWith((message) => updates(
-          message as CharacteristicAddress)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as CharacteristicAddress))
+          as CharacteristicAddress; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static CharacteristicAddress create() => CharacteristicAddress._();
@@ -1449,7 +1702,7 @@ class CharacteristicAddress extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static CharacteristicAddress getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<CharacteristicAddress>(create);
-  static CharacteristicAddress _defaultInstance;
+  static CharacteristicAddress? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.String get deviceId => $_getSZ(0);
@@ -1514,7 +1767,19 @@ class ServiceDataEntry extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   ServiceDataEntry._() : super();
-  factory ServiceDataEntry() => create();
+  factory ServiceDataEntry({
+    Uuid? serviceUuid,
+    $core.List<$core.int>? data,
+  }) {
+    final _result = create();
+    if (serviceUuid != null) {
+      _result.serviceUuid = serviceUuid;
+    }
+    if (data != null) {
+      _result.data = data;
+    }
+    return _result;
+  }
   factory ServiceDataEntry.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1529,8 +1794,8 @@ class ServiceDataEntry extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   ServiceDataEntry copyWith(void Function(ServiceDataEntry) updates) =>
-      super.copyWith((message) => updates(
-          message as ServiceDataEntry)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as ServiceDataEntry))
+          as ServiceDataEntry; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static ServiceDataEntry create() => ServiceDataEntry._();
@@ -1540,7 +1805,7 @@ class ServiceDataEntry extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static ServiceDataEntry getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<ServiceDataEntry>(create);
-  static ServiceDataEntry _defaultInstance;
+  static ServiceDataEntry? _defaultInstance;
 
   @$pb.TagNumber(1)
   Uuid get serviceUuid => $_getN(0);
@@ -1585,7 +1850,15 @@ class ServicesWithCharacteristics extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   ServicesWithCharacteristics._() : super();
-  factory ServicesWithCharacteristics() => create();
+  factory ServicesWithCharacteristics({
+    $core.Iterable<ServiceWithCharacteristics>? items,
+  }) {
+    final _result = create();
+    if (items != null) {
+      _result.items.addAll(items);
+    }
+    return _result;
+  }
   factory ServicesWithCharacteristics.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1602,8 +1875,9 @@ class ServicesWithCharacteristics extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   ServicesWithCharacteristics copyWith(
           void Function(ServicesWithCharacteristics) updates) =>
-      super.copyWith((message) => updates(message
-          as ServicesWithCharacteristics)); // ignore: deprecated_member_use
+      super.copyWith(
+              (message) => updates(message as ServicesWithCharacteristics))
+          as ServicesWithCharacteristics; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static ServicesWithCharacteristics create() =>
@@ -1614,7 +1888,7 @@ class ServicesWithCharacteristics extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static ServicesWithCharacteristics getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<ServicesWithCharacteristics>(create);
-  static ServicesWithCharacteristics _defaultInstance;
+  static ServicesWithCharacteristics? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.List<ServiceWithCharacteristics> get items => $_getList(0);
@@ -1643,7 +1917,19 @@ class ServiceWithCharacteristics extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   ServiceWithCharacteristics._() : super();
-  factory ServiceWithCharacteristics() => create();
+  factory ServiceWithCharacteristics({
+    Uuid? serviceId,
+    $core.Iterable<Uuid>? characteristics,
+  }) {
+    final _result = create();
+    if (serviceId != null) {
+      _result.serviceId = serviceId;
+    }
+    if (characteristics != null) {
+      _result.characteristics.addAll(characteristics);
+    }
+    return _result;
+  }
   factory ServiceWithCharacteristics.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1660,8 +1946,9 @@ class ServiceWithCharacteristics extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   ServiceWithCharacteristics copyWith(
           void Function(ServiceWithCharacteristics) updates) =>
-      super.copyWith((message) => updates(message
-          as ServiceWithCharacteristics)); // ignore: deprecated_member_use
+      super.copyWith(
+              (message) => updates(message as ServiceWithCharacteristics))
+          as ServiceWithCharacteristics; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static ServiceWithCharacteristics create() => ServiceWithCharacteristics._();
@@ -1671,7 +1958,7 @@ class ServiceWithCharacteristics extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static ServiceWithCharacteristics getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<ServiceWithCharacteristics>(create);
-  static ServiceWithCharacteristics _defaultInstance;
+  static ServiceWithCharacteristics? _defaultInstance;
 
   @$pb.TagNumber(1)
   Uuid get serviceId => $_getN(0);
@@ -1706,7 +1993,15 @@ class DiscoverServicesRequest extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   DiscoverServicesRequest._() : super();
-  factory DiscoverServicesRequest() => create();
+  factory DiscoverServicesRequest({
+    $core.String? deviceId,
+  }) {
+    final _result = create();
+    if (deviceId != null) {
+      _result.deviceId = deviceId;
+    }
+    return _result;
+  }
   factory DiscoverServicesRequest.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1723,8 +2018,8 @@ class DiscoverServicesRequest extends $pb.GeneratedMessage {
       'Will be removed in next major version')
   DiscoverServicesRequest copyWith(
           void Function(DiscoverServicesRequest) updates) =>
-      super.copyWith((message) => updates(
-          message as DiscoverServicesRequest)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as DiscoverServicesRequest))
+          as DiscoverServicesRequest; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static DiscoverServicesRequest create() => DiscoverServicesRequest._();
@@ -1734,7 +2029,7 @@ class DiscoverServicesRequest extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static DiscoverServicesRequest getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<DiscoverServicesRequest>(create);
-  static DiscoverServicesRequest _defaultInstance;
+  static DiscoverServicesRequest? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.String get deviceId => $_getSZ(0);
@@ -1772,7 +2067,19 @@ class DiscoverServicesInfo extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   DiscoverServicesInfo._() : super();
-  factory DiscoverServicesInfo() => create();
+  factory DiscoverServicesInfo({
+    $core.String? deviceId,
+    $core.Iterable<DiscoveredService>? services,
+  }) {
+    final _result = create();
+    if (deviceId != null) {
+      _result.deviceId = deviceId;
+    }
+    if (services != null) {
+      _result.services.addAll(services);
+    }
+    return _result;
+  }
   factory DiscoverServicesInfo.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1788,8 +2095,8 @@ class DiscoverServicesInfo extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   DiscoverServicesInfo copyWith(void Function(DiscoverServicesInfo) updates) =>
-      super.copyWith((message) => updates(
-          message as DiscoverServicesInfo)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as DiscoverServicesInfo))
+          as DiscoverServicesInfo; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static DiscoverServicesInfo create() => DiscoverServicesInfo._();
@@ -1799,7 +2106,7 @@ class DiscoverServicesInfo extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static DiscoverServicesInfo getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<DiscoverServicesInfo>(create);
-  static DiscoverServicesInfo _defaultInstance;
+  static DiscoverServicesInfo? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.String get deviceId => $_getSZ(0);
@@ -1842,7 +2149,23 @@ class DiscoveredService extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   DiscoveredService._() : super();
-  factory DiscoveredService() => create();
+  factory DiscoveredService({
+    Uuid? serviceUuid,
+    $core.Iterable<Uuid>? characteristicUuids,
+    $core.Iterable<DiscoveredService>? includedServices,
+  }) {
+    final _result = create();
+    if (serviceUuid != null) {
+      _result.serviceUuid = serviceUuid;
+    }
+    if (characteristicUuids != null) {
+      _result.characteristicUuids.addAll(characteristicUuids);
+    }
+    if (includedServices != null) {
+      _result.includedServices.addAll(includedServices);
+    }
+    return _result;
+  }
   factory DiscoveredService.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1857,8 +2180,8 @@ class DiscoveredService extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   DiscoveredService copyWith(void Function(DiscoveredService) updates) =>
-      super.copyWith((message) => updates(
-          message as DiscoveredService)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as DiscoveredService))
+          as DiscoveredService; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static DiscoveredService create() => DiscoveredService._();
@@ -1868,7 +2191,7 @@ class DiscoveredService extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static DiscoveredService getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<DiscoveredService>(create);
-  static DiscoveredService _defaultInstance;
+  static DiscoveredService? _defaultInstance;
 
   @$pb.TagNumber(1)
   Uuid get serviceUuid => $_getN(0);
@@ -1906,7 +2229,15 @@ class Uuid extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   Uuid._() : super();
-  factory Uuid() => create();
+  factory Uuid({
+    $core.List<$core.int>? data,
+  }) {
+    final _result = create();
+    if (data != null) {
+      _result.data = data;
+    }
+    return _result;
+  }
   factory Uuid.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1920,8 +2251,9 @@ class Uuid extends $pb.GeneratedMessage {
   @$core.Deprecated('Using this can add significant overhead to your binary. '
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
-  Uuid copyWith(void Function(Uuid) updates) => super.copyWith(
-      (message) => updates(message as Uuid)); // ignore: deprecated_member_use
+  Uuid copyWith(void Function(Uuid) updates) =>
+      super.copyWith((message) => updates(message as Uuid))
+          as Uuid; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static Uuid create() => Uuid._();
@@ -1930,7 +2262,7 @@ class Uuid extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static Uuid getDefault() =>
       _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<Uuid>(create);
-  static Uuid _defaultInstance;
+  static Uuid? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.List<$core.int> get data => $_getN(0);
@@ -1965,7 +2297,19 @@ class GenericFailure extends $pb.GeneratedMessage {
     ..hasRequiredFields = false;
 
   GenericFailure._() : super();
-  factory GenericFailure() => create();
+  factory GenericFailure({
+    $core.int? code,
+    $core.String? message,
+  }) {
+    final _result = create();
+    if (code != null) {
+      _result.code = code;
+    }
+    if (message != null) {
+      _result.message = message;
+    }
+    return _result;
+  }
   factory GenericFailure.fromBuffer($core.List<$core.int> i,
           [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) =>
       create()..mergeFromBuffer(i, r);
@@ -1980,8 +2324,8 @@ class GenericFailure extends $pb.GeneratedMessage {
       'Use [GeneratedMessageGenericExtensions.rebuild] instead. '
       'Will be removed in next major version')
   GenericFailure copyWith(void Function(GenericFailure) updates) =>
-      super.copyWith((message) =>
-          updates(message as GenericFailure)); // ignore: deprecated_member_use
+      super.copyWith((message) => updates(message as GenericFailure))
+          as GenericFailure; // ignore: deprecated_member_use
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static GenericFailure create() => GenericFailure._();
@@ -1991,7 +2335,7 @@ class GenericFailure extends $pb.GeneratedMessage {
   @$core.pragma('dart2js:noInline')
   static GenericFailure getDefault() => _defaultInstance ??=
       $pb.GeneratedMessage.$_defaultFor<GenericFailure>(create);
-  static GenericFailure _defaultInstance;
+  static GenericFailure? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.int get code => $_getIZ(0);
