@@ -3,7 +3,7 @@ package com.signify.hue.flutterreactiveble.ble
 import com.polidea.rxandroidble2.RxBleConnection
 import java.util.UUID
 
-data class ScanInfo(val deviceId: String, val name: String, val rssi: Int, val serviceData: Map<UUID, ByteArray>, val manufacturerData: ByteArray) {
+data class ScanInfo(val deviceId: String, val name: String, val rssi: Int, val serviceData: Map<UUID, ByteArray>, val serviceUuids: List<UUID>, val manufacturerData: ByteArray) {
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (javaClass != other?.javaClass) return false
@@ -14,6 +14,7 @@ data class ScanInfo(val deviceId: String, val name: String, val rssi: Int, val s
         if (name != other.name) return false
         if (rssi != other.rssi) return false
         if (serviceData != other.serviceData) return false
+        if (serviceUuids != other.serviceUuids) return false
         if (!manufacturerData.contentEquals(other.manufacturerData)) return false
 
         return true
@@ -24,6 +25,7 @@ data class ScanInfo(val deviceId: String, val name: String, val rssi: Int, val s
         result = 31 * result + name.hashCode()
         result = 31 * result + rssi
         result = 31 * result + serviceData.hashCode()
+        result = 31 * result + serviceUuids.hashCode()
         result = 31 * result + manufacturerData.contentHashCode()
         return result
     }
