@@ -1,5 +1,4 @@
 // DO NOT EDIT.
-// swift-format-ignore-file
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
 // Source: bledata.proto
@@ -13,7 +12,7 @@ import SwiftProtobuf
 // If the compiler emits an error on this type, it is because this file
 // was generated by a version of the `protoc` Swift plug-in that is
 // incompatible with the version of SwiftProtobuf to which you are linking.
-// Please ensure that you are building against the same version of the API
+// Please ensure that your are building against the same version of the API
 // that was used to generate this file.
 fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAPIVersionCheck {
   struct _2: SwiftProtobuf.ProtobufAPIVersion_2 {}
@@ -41,30 +40,50 @@ struct DeviceScanInfo {
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
 
-  var id: String = String()
+  var id: String {
+    get {return _storage._id}
+    set {_uniqueStorage()._id = newValue}
+  }
 
-  var name: String = String()
+  var name: String {
+    get {return _storage._name}
+    set {_uniqueStorage()._name = newValue}
+  }
 
   var failure: GenericFailure {
-    get {return _failure ?? GenericFailure()}
-    set {_failure = newValue}
+    get {return _storage._failure ?? GenericFailure()}
+    set {_uniqueStorage()._failure = newValue}
   }
   /// Returns true if `failure` has been explicitly set.
-  var hasFailure: Bool {return self._failure != nil}
+  var hasFailure: Bool {return _storage._failure != nil}
   /// Clears the value of `failure`. Subsequent reads from it will return its default value.
-  mutating func clearFailure() {self._failure = nil}
+  mutating func clearFailure() {_uniqueStorage()._failure = nil}
 
-  var serviceData: [ServiceDataEntry] = []
+  var serviceData: [ServiceDataEntry] {
+    get {return _storage._serviceData}
+    set {_uniqueStorage()._serviceData = newValue}
+  }
 
-  var manufacturerData: Data = Data()
+  var serviceUuids: [Uuid] {
+    get {return _storage._serviceUuids}
+    set {_uniqueStorage()._serviceUuids = newValue}
+  }
 
-  var rssi: Int32 = 0
+  var manufacturerData: Data {
+    get {return _storage._manufacturerData}
+    set {_uniqueStorage()._manufacturerData = newValue}
+  }
+
+  var rssi: Int32 {
+    get {return _storage._rssi}
+    set {_uniqueStorage()._rssi = newValue}
+  }
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _failure: GenericFailure? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct ConnectToDeviceRequest {
@@ -72,24 +91,30 @@ struct ConnectToDeviceRequest {
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
 
-  var deviceID: String = String()
+  var deviceID: String {
+    get {return _storage._deviceID}
+    set {_uniqueStorage()._deviceID = newValue}
+  }
 
   var servicesWithCharacteristicsToDiscover: ServicesWithCharacteristics {
-    get {return _servicesWithCharacteristicsToDiscover ?? ServicesWithCharacteristics()}
-    set {_servicesWithCharacteristicsToDiscover = newValue}
+    get {return _storage._servicesWithCharacteristicsToDiscover ?? ServicesWithCharacteristics()}
+    set {_uniqueStorage()._servicesWithCharacteristicsToDiscover = newValue}
   }
   /// Returns true if `servicesWithCharacteristicsToDiscover` has been explicitly set.
-  var hasServicesWithCharacteristicsToDiscover: Bool {return self._servicesWithCharacteristicsToDiscover != nil}
+  var hasServicesWithCharacteristicsToDiscover: Bool {return _storage._servicesWithCharacteristicsToDiscover != nil}
   /// Clears the value of `servicesWithCharacteristicsToDiscover`. Subsequent reads from it will return its default value.
-  mutating func clearServicesWithCharacteristicsToDiscover() {self._servicesWithCharacteristicsToDiscover = nil}
+  mutating func clearServicesWithCharacteristicsToDiscover() {_uniqueStorage()._servicesWithCharacteristicsToDiscover = nil}
 
-  var timeoutInMs: Int32 = 0
+  var timeoutInMs: Int32 {
+    get {return _storage._timeoutInMs}
+    set {_uniqueStorage()._timeoutInMs = newValue}
+  }
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _servicesWithCharacteristicsToDiscover: ServicesWithCharacteristics? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct DeviceInfo {
@@ -97,24 +122,30 @@ struct DeviceInfo {
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
 
-  var id: String = String()
+  var id: String {
+    get {return _storage._id}
+    set {_uniqueStorage()._id = newValue}
+  }
 
-  var connectionState: Int32 = 0
+  var connectionState: Int32 {
+    get {return _storage._connectionState}
+    set {_uniqueStorage()._connectionState = newValue}
+  }
 
   var failure: GenericFailure {
-    get {return _failure ?? GenericFailure()}
-    set {_failure = newValue}
+    get {return _storage._failure ?? GenericFailure()}
+    set {_uniqueStorage()._failure = newValue}
   }
   /// Returns true if `failure` has been explicitly set.
-  var hasFailure: Bool {return self._failure != nil}
+  var hasFailure: Bool {return _storage._failure != nil}
   /// Clears the value of `failure`. Subsequent reads from it will return its default value.
-  mutating func clearFailure() {self._failure = nil}
+  mutating func clearFailure() {_uniqueStorage()._failure = nil}
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _failure: GenericFailure? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct DisconnectFromDeviceRequest {
@@ -147,19 +178,19 @@ struct ClearGattCacheInfo {
   // methods supported on all messages.
 
   var failure: GenericFailure {
-    get {return _failure ?? GenericFailure()}
-    set {_failure = newValue}
+    get {return _storage._failure ?? GenericFailure()}
+    set {_uniqueStorage()._failure = newValue}
   }
   /// Returns true if `failure` has been explicitly set.
-  var hasFailure: Bool {return self._failure != nil}
+  var hasFailure: Bool {return _storage._failure != nil}
   /// Clears the value of `failure`. Subsequent reads from it will return its default value.
-  mutating func clearFailure() {self._failure = nil}
+  mutating func clearFailure() {_uniqueStorage()._failure = nil}
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _failure: GenericFailure? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct NotifyCharacteristicRequest {
@@ -168,19 +199,19 @@ struct NotifyCharacteristicRequest {
   // methods supported on all messages.
 
   var characteristic: CharacteristicAddress {
-    get {return _characteristic ?? CharacteristicAddress()}
-    set {_characteristic = newValue}
+    get {return _storage._characteristic ?? CharacteristicAddress()}
+    set {_uniqueStorage()._characteristic = newValue}
   }
   /// Returns true if `characteristic` has been explicitly set.
-  var hasCharacteristic: Bool {return self._characteristic != nil}
+  var hasCharacteristic: Bool {return _storage._characteristic != nil}
   /// Clears the value of `characteristic`. Subsequent reads from it will return its default value.
-  mutating func clearCharacteristic() {self._characteristic = nil}
+  mutating func clearCharacteristic() {_uniqueStorage()._characteristic = nil}
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _characteristic: CharacteristicAddress? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct NotifyNoMoreCharacteristicRequest {
@@ -189,19 +220,19 @@ struct NotifyNoMoreCharacteristicRequest {
   // methods supported on all messages.
 
   var characteristic: CharacteristicAddress {
-    get {return _characteristic ?? CharacteristicAddress()}
-    set {_characteristic = newValue}
+    get {return _storage._characteristic ?? CharacteristicAddress()}
+    set {_uniqueStorage()._characteristic = newValue}
   }
   /// Returns true if `characteristic` has been explicitly set.
-  var hasCharacteristic: Bool {return self._characteristic != nil}
+  var hasCharacteristic: Bool {return _storage._characteristic != nil}
   /// Clears the value of `characteristic`. Subsequent reads from it will return its default value.
-  mutating func clearCharacteristic() {self._characteristic = nil}
+  mutating func clearCharacteristic() {_uniqueStorage()._characteristic = nil}
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _characteristic: CharacteristicAddress? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct ReadCharacteristicRequest {
@@ -210,19 +241,19 @@ struct ReadCharacteristicRequest {
   // methods supported on all messages.
 
   var characteristic: CharacteristicAddress {
-    get {return _characteristic ?? CharacteristicAddress()}
-    set {_characteristic = newValue}
+    get {return _storage._characteristic ?? CharacteristicAddress()}
+    set {_uniqueStorage()._characteristic = newValue}
   }
   /// Returns true if `characteristic` has been explicitly set.
-  var hasCharacteristic: Bool {return self._characteristic != nil}
+  var hasCharacteristic: Bool {return _storage._characteristic != nil}
   /// Clears the value of `characteristic`. Subsequent reads from it will return its default value.
-  mutating func clearCharacteristic() {self._characteristic = nil}
+  mutating func clearCharacteristic() {_uniqueStorage()._characteristic = nil}
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _characteristic: CharacteristicAddress? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct CharacteristicValueInfo {
@@ -231,31 +262,33 @@ struct CharacteristicValueInfo {
   // methods supported on all messages.
 
   var characteristic: CharacteristicAddress {
-    get {return _characteristic ?? CharacteristicAddress()}
-    set {_characteristic = newValue}
+    get {return _storage._characteristic ?? CharacteristicAddress()}
+    set {_uniqueStorage()._characteristic = newValue}
   }
   /// Returns true if `characteristic` has been explicitly set.
-  var hasCharacteristic: Bool {return self._characteristic != nil}
+  var hasCharacteristic: Bool {return _storage._characteristic != nil}
   /// Clears the value of `characteristic`. Subsequent reads from it will return its default value.
-  mutating func clearCharacteristic() {self._characteristic = nil}
+  mutating func clearCharacteristic() {_uniqueStorage()._characteristic = nil}
 
-  var value: Data = Data()
+  var value: Data {
+    get {return _storage._value}
+    set {_uniqueStorage()._value = newValue}
+  }
 
   var failure: GenericFailure {
-    get {return _failure ?? GenericFailure()}
-    set {_failure = newValue}
+    get {return _storage._failure ?? GenericFailure()}
+    set {_uniqueStorage()._failure = newValue}
   }
   /// Returns true if `failure` has been explicitly set.
-  var hasFailure: Bool {return self._failure != nil}
+  var hasFailure: Bool {return _storage._failure != nil}
   /// Clears the value of `failure`. Subsequent reads from it will return its default value.
-  mutating func clearFailure() {self._failure = nil}
+  mutating func clearFailure() {_uniqueStorage()._failure = nil}
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _characteristic: CharacteristicAddress? = nil
-  fileprivate var _failure: GenericFailure? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct WriteCharacteristicRequest {
@@ -264,21 +297,24 @@ struct WriteCharacteristicRequest {
   // methods supported on all messages.
 
   var characteristic: CharacteristicAddress {
-    get {return _characteristic ?? CharacteristicAddress()}
-    set {_characteristic = newValue}
+    get {return _storage._characteristic ?? CharacteristicAddress()}
+    set {_uniqueStorage()._characteristic = newValue}
   }
   /// Returns true if `characteristic` has been explicitly set.
-  var hasCharacteristic: Bool {return self._characteristic != nil}
+  var hasCharacteristic: Bool {return _storage._characteristic != nil}
   /// Clears the value of `characteristic`. Subsequent reads from it will return its default value.
-  mutating func clearCharacteristic() {self._characteristic = nil}
+  mutating func clearCharacteristic() {_uniqueStorage()._characteristic = nil}
 
-  var value: Data = Data()
+  var value: Data {
+    get {return _storage._value}
+    set {_uniqueStorage()._value = newValue}
+  }
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _characteristic: CharacteristicAddress? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct WriteCharacteristicInfo {
@@ -287,29 +323,28 @@ struct WriteCharacteristicInfo {
   // methods supported on all messages.
 
   var characteristic: CharacteristicAddress {
-    get {return _characteristic ?? CharacteristicAddress()}
-    set {_characteristic = newValue}
+    get {return _storage._characteristic ?? CharacteristicAddress()}
+    set {_uniqueStorage()._characteristic = newValue}
   }
   /// Returns true if `characteristic` has been explicitly set.
-  var hasCharacteristic: Bool {return self._characteristic != nil}
+  var hasCharacteristic: Bool {return _storage._characteristic != nil}
   /// Clears the value of `characteristic`. Subsequent reads from it will return its default value.
-  mutating func clearCharacteristic() {self._characteristic = nil}
+  mutating func clearCharacteristic() {_uniqueStorage()._characteristic = nil}
 
   var failure: GenericFailure {
-    get {return _failure ?? GenericFailure()}
-    set {_failure = newValue}
+    get {return _storage._failure ?? GenericFailure()}
+    set {_uniqueStorage()._failure = newValue}
   }
   /// Returns true if `failure` has been explicitly set.
-  var hasFailure: Bool {return self._failure != nil}
+  var hasFailure: Bool {return _storage._failure != nil}
   /// Clears the value of `failure`. Subsequent reads from it will return its default value.
-  mutating func clearFailure() {self._failure = nil}
+  mutating func clearFailure() {_uniqueStorage()._failure = nil}
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _characteristic: CharacteristicAddress? = nil
-  fileprivate var _failure: GenericFailure? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct NegotiateMtuRequest {
@@ -331,24 +366,30 @@ struct NegotiateMtuInfo {
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
 
-  var deviceID: String = String()
+  var deviceID: String {
+    get {return _storage._deviceID}
+    set {_uniqueStorage()._deviceID = newValue}
+  }
 
-  var mtuSize: Int32 = 0
+  var mtuSize: Int32 {
+    get {return _storage._mtuSize}
+    set {_uniqueStorage()._mtuSize = newValue}
+  }
 
   var failure: GenericFailure {
-    get {return _failure ?? GenericFailure()}
-    set {_failure = newValue}
+    get {return _storage._failure ?? GenericFailure()}
+    set {_uniqueStorage()._failure = newValue}
   }
   /// Returns true if `failure` has been explicitly set.
-  var hasFailure: Bool {return self._failure != nil}
+  var hasFailure: Bool {return _storage._failure != nil}
   /// Clears the value of `failure`. Subsequent reads from it will return its default value.
-  mutating func clearFailure() {self._failure = nil}
+  mutating func clearFailure() {_uniqueStorage()._failure = nil}
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _failure: GenericFailure? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct BleStatusInfo {
@@ -382,22 +423,25 @@ struct ChangeConnectionPriorityInfo {
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
 
-  var deviceID: String = String()
+  var deviceID: String {
+    get {return _storage._deviceID}
+    set {_uniqueStorage()._deviceID = newValue}
+  }
 
   var failure: GenericFailure {
-    get {return _failure ?? GenericFailure()}
-    set {_failure = newValue}
+    get {return _storage._failure ?? GenericFailure()}
+    set {_uniqueStorage()._failure = newValue}
   }
   /// Returns true if `failure` has been explicitly set.
-  var hasFailure: Bool {return self._failure != nil}
+  var hasFailure: Bool {return _storage._failure != nil}
   /// Clears the value of `failure`. Subsequent reads from it will return its default value.
-  mutating func clearFailure() {self._failure = nil}
+  mutating func clearFailure() {_uniqueStorage()._failure = nil}
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _failure: GenericFailure? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct CharacteristicAddress {
@@ -405,32 +449,34 @@ struct CharacteristicAddress {
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
 
-  var deviceID: String = String()
+  var deviceID: String {
+    get {return _storage._deviceID}
+    set {_uniqueStorage()._deviceID = newValue}
+  }
 
   var serviceUuid: Uuid {
-    get {return _serviceUuid ?? Uuid()}
-    set {_serviceUuid = newValue}
+    get {return _storage._serviceUuid ?? Uuid()}
+    set {_uniqueStorage()._serviceUuid = newValue}
   }
   /// Returns true if `serviceUuid` has been explicitly set.
-  var hasServiceUuid: Bool {return self._serviceUuid != nil}
+  var hasServiceUuid: Bool {return _storage._serviceUuid != nil}
   /// Clears the value of `serviceUuid`. Subsequent reads from it will return its default value.
-  mutating func clearServiceUuid() {self._serviceUuid = nil}
+  mutating func clearServiceUuid() {_uniqueStorage()._serviceUuid = nil}
 
   var characteristicUuid: Uuid {
-    get {return _characteristicUuid ?? Uuid()}
-    set {_characteristicUuid = newValue}
+    get {return _storage._characteristicUuid ?? Uuid()}
+    set {_uniqueStorage()._characteristicUuid = newValue}
   }
   /// Returns true if `characteristicUuid` has been explicitly set.
-  var hasCharacteristicUuid: Bool {return self._characteristicUuid != nil}
+  var hasCharacteristicUuid: Bool {return _storage._characteristicUuid != nil}
   /// Clears the value of `characteristicUuid`. Subsequent reads from it will return its default value.
-  mutating func clearCharacteristicUuid() {self._characteristicUuid = nil}
+  mutating func clearCharacteristicUuid() {_uniqueStorage()._characteristicUuid = nil}
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _serviceUuid: Uuid? = nil
-  fileprivate var _characteristicUuid: Uuid? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct ServiceDataEntry {
@@ -439,21 +485,24 @@ struct ServiceDataEntry {
   // methods supported on all messages.
 
   var serviceUuid: Uuid {
-    get {return _serviceUuid ?? Uuid()}
-    set {_serviceUuid = newValue}
+    get {return _storage._serviceUuid ?? Uuid()}
+    set {_uniqueStorage()._serviceUuid = newValue}
   }
   /// Returns true if `serviceUuid` has been explicitly set.
-  var hasServiceUuid: Bool {return self._serviceUuid != nil}
+  var hasServiceUuid: Bool {return _storage._serviceUuid != nil}
   /// Clears the value of `serviceUuid`. Subsequent reads from it will return its default value.
-  mutating func clearServiceUuid() {self._serviceUuid = nil}
+  mutating func clearServiceUuid() {_uniqueStorage()._serviceUuid = nil}
 
-  var data: Data = Data()
+  var data: Data {
+    get {return _storage._data}
+    set {_uniqueStorage()._data = newValue}
+  }
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _serviceUuid: Uuid? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct ServicesWithCharacteristics {
@@ -474,21 +523,24 @@ struct ServiceWithCharacteristics {
   // methods supported on all messages.
 
   var serviceID: Uuid {
-    get {return _serviceID ?? Uuid()}
-    set {_serviceID = newValue}
+    get {return _storage._serviceID ?? Uuid()}
+    set {_uniqueStorage()._serviceID = newValue}
   }
   /// Returns true if `serviceID` has been explicitly set.
-  var hasServiceID: Bool {return self._serviceID != nil}
+  var hasServiceID: Bool {return _storage._serviceID != nil}
   /// Clears the value of `serviceID`. Subsequent reads from it will return its default value.
-  mutating func clearServiceID() {self._serviceID = nil}
+  mutating func clearServiceID() {_uniqueStorage()._serviceID = nil}
 
-  var characteristics: [Uuid] = []
+  var characteristics: [Uuid] {
+    get {return _storage._characteristics}
+    set {_uniqueStorage()._characteristics = newValue}
+  }
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _serviceID: Uuid? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct DiscoverServicesRequest {
@@ -523,23 +575,29 @@ struct DiscoveredService {
   // methods supported on all messages.
 
   var serviceUuid: Uuid {
-    get {return _serviceUuid ?? Uuid()}
-    set {_serviceUuid = newValue}
+    get {return _storage._serviceUuid ?? Uuid()}
+    set {_uniqueStorage()._serviceUuid = newValue}
   }
   /// Returns true if `serviceUuid` has been explicitly set.
-  var hasServiceUuid: Bool {return self._serviceUuid != nil}
+  var hasServiceUuid: Bool {return _storage._serviceUuid != nil}
   /// Clears the value of `serviceUuid`. Subsequent reads from it will return its default value.
-  mutating func clearServiceUuid() {self._serviceUuid = nil}
+  mutating func clearServiceUuid() {_uniqueStorage()._serviceUuid = nil}
 
-  var characteristicUuids: [Uuid] = []
+  var characteristicUuids: [Uuid] {
+    get {return _storage._characteristicUuids}
+    set {_uniqueStorage()._characteristicUuids = newValue}
+  }
 
-  var includedServices: [DiscoveredService] = []
+  var includedServices: [DiscoveredService] {
+    get {return _storage._includedServices}
+    set {_uniqueStorage()._includedServices = newValue}
+  }
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
   init() {}
 
-  fileprivate var _serviceUuid: Uuid? = nil
+  fileprivate var _storage = _StorageClass.defaultInstance
 }
 
 struct Uuid {
@@ -547,7 +605,7 @@ struct Uuid {
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
 
-  var data: Data = Data()
+  var data: Data = SwiftProtobuf.Internal.emptyData
 
   var unknownFields = SwiftProtobuf.UnknownStorage()
 
@@ -580,13 +638,10 @@ extension ScanForDevicesRequest: SwiftProtobuf.Message, SwiftProtobuf._MessageIm
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try { try decoder.decodeRepeatedMessageField(value: &self.serviceUuids) }()
-      case 2: try { try decoder.decodeSingularInt32Field(value: &self.scanMode) }()
-      case 3: try { try decoder.decodeSingularBoolField(value: &self.requireLocationServicesEnabled) }()
+      case 1: try decoder.decodeRepeatedMessageField(value: &self.serviceUuids)
+      case 2: try decoder.decodeSingularInt32Field(value: &self.scanMode)
+      case 3: try decoder.decodeSingularBoolField(value: &self.requireLocationServicesEnabled)
       default: break
       }
     }
@@ -621,56 +676,103 @@ extension DeviceScanInfo: SwiftProtobuf.Message, SwiftProtobuf._MessageImplement
     2: .same(proto: "name"),
     3: .same(proto: "failure"),
     4: .same(proto: "serviceData"),
+    7: .same(proto: "serviceUuids"),
     6: .same(proto: "manufacturerData"),
     5: .same(proto: "rssi"),
   ]
 
+  fileprivate class _StorageClass {
+    var _id: String = String()
+    var _name: String = String()
+    var _failure: GenericFailure? = nil
+    var _serviceData: [ServiceDataEntry] = []
+    var _serviceUuids: [Uuid] = []
+    var _manufacturerData: Data = SwiftProtobuf.Internal.emptyData
+    var _rssi: Int32 = 0
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _id = source._id
+      _name = source._name
+      _failure = source._failure
+      _serviceData = source._serviceData
+      _serviceUuids = source._serviceUuids
+      _manufacturerData = source._manufacturerData
+      _rssi = source._rssi
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.id) }()
-      case 2: try { try decoder.decodeSingularStringField(value: &self.name) }()
-      case 3: try { try decoder.decodeSingularMessageField(value: &self._failure) }()
-      case 4: try { try decoder.decodeRepeatedMessageField(value: &self.serviceData) }()
-      case 5: try { try decoder.decodeSingularInt32Field(value: &self.rssi) }()
-      case 6: try { try decoder.decodeSingularBytesField(value: &self.manufacturerData) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularStringField(value: &_storage._id)
+        case 2: try decoder.decodeSingularStringField(value: &_storage._name)
+        case 3: try decoder.decodeSingularMessageField(value: &_storage._failure)
+        case 4: try decoder.decodeRepeatedMessageField(value: &_storage._serviceData)
+        case 5: try decoder.decodeSingularInt32Field(value: &_storage._rssi)
+        case 6: try decoder.decodeSingularBytesField(value: &_storage._manufacturerData)
+        case 7: try decoder.decodeRepeatedMessageField(value: &_storage._serviceUuids)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if !self.id.isEmpty {
-      try visitor.visitSingularStringField(value: self.id, fieldNumber: 1)
-    }
-    if !self.name.isEmpty {
-      try visitor.visitSingularStringField(value: self.name, fieldNumber: 2)
-    }
-    if let v = self._failure {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
-    }
-    if !self.serviceData.isEmpty {
-      try visitor.visitRepeatedMessageField(value: self.serviceData, fieldNumber: 4)
-    }
-    if self.rssi != 0 {
-      try visitor.visitSingularInt32Field(value: self.rssi, fieldNumber: 5)
-    }
-    if !self.manufacturerData.isEmpty {
-      try visitor.visitSingularBytesField(value: self.manufacturerData, fieldNumber: 6)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if !_storage._id.isEmpty {
+        try visitor.visitSingularStringField(value: _storage._id, fieldNumber: 1)
+      }
+      if !_storage._name.isEmpty {
+        try visitor.visitSingularStringField(value: _storage._name, fieldNumber: 2)
+      }
+      if let v = _storage._failure {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
+      }
+      if !_storage._serviceData.isEmpty {
+        try visitor.visitRepeatedMessageField(value: _storage._serviceData, fieldNumber: 4)
+      }
+      if _storage._rssi != 0 {
+        try visitor.visitSingularInt32Field(value: _storage._rssi, fieldNumber: 5)
+      }
+      if !_storage._manufacturerData.isEmpty {
+        try visitor.visitSingularBytesField(value: _storage._manufacturerData, fieldNumber: 6)
+      }
+      if !_storage._serviceUuids.isEmpty {
+        try visitor.visitRepeatedMessageField(value: _storage._serviceUuids, fieldNumber: 7)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: DeviceScanInfo, rhs: DeviceScanInfo) -> Bool {
-    if lhs.id != rhs.id {return false}
-    if lhs.name != rhs.name {return false}
-    if lhs._failure != rhs._failure {return false}
-    if lhs.serviceData != rhs.serviceData {return false}
-    if lhs.manufacturerData != rhs.manufacturerData {return false}
-    if lhs.rssi != rhs.rssi {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._id != rhs_storage._id {return false}
+        if _storage._name != rhs_storage._name {return false}
+        if _storage._failure != rhs_storage._failure {return false}
+        if _storage._serviceData != rhs_storage._serviceData {return false}
+        if _storage._serviceUuids != rhs_storage._serviceUuids {return false}
+        if _storage._manufacturerData != rhs_storage._manufacturerData {return false}
+        if _storage._rssi != rhs_storage._rssi {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -684,37 +786,70 @@ extension ConnectToDeviceRequest: SwiftProtobuf.Message, SwiftProtobuf._MessageI
     3: .same(proto: "timeoutInMs"),
   ]
 
+  fileprivate class _StorageClass {
+    var _deviceID: String = String()
+    var _servicesWithCharacteristicsToDiscover: ServicesWithCharacteristics? = nil
+    var _timeoutInMs: Int32 = 0
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _deviceID = source._deviceID
+      _servicesWithCharacteristicsToDiscover = source._servicesWithCharacteristicsToDiscover
+      _timeoutInMs = source._timeoutInMs
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.deviceID) }()
-      case 2: try { try decoder.decodeSingularMessageField(value: &self._servicesWithCharacteristicsToDiscover) }()
-      case 3: try { try decoder.decodeSingularInt32Field(value: &self.timeoutInMs) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularStringField(value: &_storage._deviceID)
+        case 2: try decoder.decodeSingularMessageField(value: &_storage._servicesWithCharacteristicsToDiscover)
+        case 3: try decoder.decodeSingularInt32Field(value: &_storage._timeoutInMs)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if !self.deviceID.isEmpty {
-      try visitor.visitSingularStringField(value: self.deviceID, fieldNumber: 1)
-    }
-    if let v = self._servicesWithCharacteristicsToDiscover {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 2)
-    }
-    if self.timeoutInMs != 0 {
-      try visitor.visitSingularInt32Field(value: self.timeoutInMs, fieldNumber: 3)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if !_storage._deviceID.isEmpty {
+        try visitor.visitSingularStringField(value: _storage._deviceID, fieldNumber: 1)
+      }
+      if let v = _storage._servicesWithCharacteristicsToDiscover {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 2)
+      }
+      if _storage._timeoutInMs != 0 {
+        try visitor.visitSingularInt32Field(value: _storage._timeoutInMs, fieldNumber: 3)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: ConnectToDeviceRequest, rhs: ConnectToDeviceRequest) -> Bool {
-    if lhs.deviceID != rhs.deviceID {return false}
-    if lhs._servicesWithCharacteristicsToDiscover != rhs._servicesWithCharacteristicsToDiscover {return false}
-    if lhs.timeoutInMs != rhs.timeoutInMs {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._deviceID != rhs_storage._deviceID {return false}
+        if _storage._servicesWithCharacteristicsToDiscover != rhs_storage._servicesWithCharacteristicsToDiscover {return false}
+        if _storage._timeoutInMs != rhs_storage._timeoutInMs {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -728,37 +863,70 @@ extension DeviceInfo: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementatio
     3: .same(proto: "failure"),
   ]
 
+  fileprivate class _StorageClass {
+    var _id: String = String()
+    var _connectionState: Int32 = 0
+    var _failure: GenericFailure? = nil
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _id = source._id
+      _connectionState = source._connectionState
+      _failure = source._failure
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.id) }()
-      case 2: try { try decoder.decodeSingularInt32Field(value: &self.connectionState) }()
-      case 3: try { try decoder.decodeSingularMessageField(value: &self._failure) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularStringField(value: &_storage._id)
+        case 2: try decoder.decodeSingularInt32Field(value: &_storage._connectionState)
+        case 3: try decoder.decodeSingularMessageField(value: &_storage._failure)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if !self.id.isEmpty {
-      try visitor.visitSingularStringField(value: self.id, fieldNumber: 1)
-    }
-    if self.connectionState != 0 {
-      try visitor.visitSingularInt32Field(value: self.connectionState, fieldNumber: 2)
-    }
-    if let v = self._failure {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if !_storage._id.isEmpty {
+        try visitor.visitSingularStringField(value: _storage._id, fieldNumber: 1)
+      }
+      if _storage._connectionState != 0 {
+        try visitor.visitSingularInt32Field(value: _storage._connectionState, fieldNumber: 2)
+      }
+      if let v = _storage._failure {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: DeviceInfo, rhs: DeviceInfo) -> Bool {
-    if lhs.id != rhs.id {return false}
-    if lhs.connectionState != rhs.connectionState {return false}
-    if lhs._failure != rhs._failure {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._id != rhs_storage._id {return false}
+        if _storage._connectionState != rhs_storage._connectionState {return false}
+        if _storage._failure != rhs_storage._failure {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -772,11 +940,8 @@ extension DisconnectFromDeviceRequest: SwiftProtobuf.Message, SwiftProtobuf._Mes
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.deviceID) }()
+      case 1: try decoder.decodeSingularStringField(value: &self.deviceID)
       default: break
       }
     }
@@ -804,11 +969,8 @@ extension ClearGattCacheRequest: SwiftProtobuf.Message, SwiftProtobuf._MessageIm
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.deviceID) }()
+      case 1: try decoder.decodeSingularStringField(value: &self.deviceID)
       default: break
       }
     }
@@ -834,27 +996,56 @@ extension ClearGattCacheInfo: SwiftProtobuf.Message, SwiftProtobuf._MessageImple
     1: .same(proto: "failure"),
   ]
 
+  fileprivate class _StorageClass {
+    var _failure: GenericFailure? = nil
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _failure = source._failure
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularMessageField(value: &self._failure) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularMessageField(value: &_storage._failure)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if let v = self._failure {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if let v = _storage._failure {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: ClearGattCacheInfo, rhs: ClearGattCacheInfo) -> Bool {
-    if lhs._failure != rhs._failure {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._failure != rhs_storage._failure {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -866,27 +1057,56 @@ extension NotifyCharacteristicRequest: SwiftProtobuf.Message, SwiftProtobuf._Mes
     1: .same(proto: "characteristic"),
   ]
 
+  fileprivate class _StorageClass {
+    var _characteristic: CharacteristicAddress? = nil
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _characteristic = source._characteristic
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularMessageField(value: &self._characteristic) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularMessageField(value: &_storage._characteristic)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if let v = self._characteristic {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if let v = _storage._characteristic {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: NotifyCharacteristicRequest, rhs: NotifyCharacteristicRequest) -> Bool {
-    if lhs._characteristic != rhs._characteristic {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._characteristic != rhs_storage._characteristic {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -898,27 +1118,56 @@ extension NotifyNoMoreCharacteristicRequest: SwiftProtobuf.Message, SwiftProtobu
     1: .same(proto: "characteristic"),
   ]
 
+  fileprivate class _StorageClass {
+    var _characteristic: CharacteristicAddress? = nil
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _characteristic = source._characteristic
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularMessageField(value: &self._characteristic) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularMessageField(value: &_storage._characteristic)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if let v = self._characteristic {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if let v = _storage._characteristic {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: NotifyNoMoreCharacteristicRequest, rhs: NotifyNoMoreCharacteristicRequest) -> Bool {
-    if lhs._characteristic != rhs._characteristic {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._characteristic != rhs_storage._characteristic {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -930,27 +1179,56 @@ extension ReadCharacteristicRequest: SwiftProtobuf.Message, SwiftProtobuf._Messa
     1: .same(proto: "characteristic"),
   ]
 
+  fileprivate class _StorageClass {
+    var _characteristic: CharacteristicAddress? = nil
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _characteristic = source._characteristic
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularMessageField(value: &self._characteristic) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularMessageField(value: &_storage._characteristic)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if let v = self._characteristic {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if let v = _storage._characteristic {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: ReadCharacteristicRequest, rhs: ReadCharacteristicRequest) -> Bool {
-    if lhs._characteristic != rhs._characteristic {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._characteristic != rhs_storage._characteristic {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -964,37 +1242,70 @@ extension CharacteristicValueInfo: SwiftProtobuf.Message, SwiftProtobuf._Message
     3: .same(proto: "failure"),
   ]
 
+  fileprivate class _StorageClass {
+    var _characteristic: CharacteristicAddress? = nil
+    var _value: Data = SwiftProtobuf.Internal.emptyData
+    var _failure: GenericFailure? = nil
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _characteristic = source._characteristic
+      _value = source._value
+      _failure = source._failure
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularMessageField(value: &self._characteristic) }()
-      case 2: try { try decoder.decodeSingularBytesField(value: &self.value) }()
-      case 3: try { try decoder.decodeSingularMessageField(value: &self._failure) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularMessageField(value: &_storage._characteristic)
+        case 2: try decoder.decodeSingularBytesField(value: &_storage._value)
+        case 3: try decoder.decodeSingularMessageField(value: &_storage._failure)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if let v = self._characteristic {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
-    }
-    if !self.value.isEmpty {
-      try visitor.visitSingularBytesField(value: self.value, fieldNumber: 2)
-    }
-    if let v = self._failure {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if let v = _storage._characteristic {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+      }
+      if !_storage._value.isEmpty {
+        try visitor.visitSingularBytesField(value: _storage._value, fieldNumber: 2)
+      }
+      if let v = _storage._failure {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: CharacteristicValueInfo, rhs: CharacteristicValueInfo) -> Bool {
-    if lhs._characteristic != rhs._characteristic {return false}
-    if lhs.value != rhs.value {return false}
-    if lhs._failure != rhs._failure {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._characteristic != rhs_storage._characteristic {return false}
+        if _storage._value != rhs_storage._value {return false}
+        if _storage._failure != rhs_storage._failure {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -1007,32 +1318,63 @@ extension WriteCharacteristicRequest: SwiftProtobuf.Message, SwiftProtobuf._Mess
     2: .same(proto: "value"),
   ]
 
+  fileprivate class _StorageClass {
+    var _characteristic: CharacteristicAddress? = nil
+    var _value: Data = SwiftProtobuf.Internal.emptyData
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _characteristic = source._characteristic
+      _value = source._value
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularMessageField(value: &self._characteristic) }()
-      case 2: try { try decoder.decodeSingularBytesField(value: &self.value) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularMessageField(value: &_storage._characteristic)
+        case 2: try decoder.decodeSingularBytesField(value: &_storage._value)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if let v = self._characteristic {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
-    }
-    if !self.value.isEmpty {
-      try visitor.visitSingularBytesField(value: self.value, fieldNumber: 2)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if let v = _storage._characteristic {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+      }
+      if !_storage._value.isEmpty {
+        try visitor.visitSingularBytesField(value: _storage._value, fieldNumber: 2)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: WriteCharacteristicRequest, rhs: WriteCharacteristicRequest) -> Bool {
-    if lhs._characteristic != rhs._characteristic {return false}
-    if lhs.value != rhs.value {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._characteristic != rhs_storage._characteristic {return false}
+        if _storage._value != rhs_storage._value {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -1045,32 +1387,63 @@ extension WriteCharacteristicInfo: SwiftProtobuf.Message, SwiftProtobuf._Message
     3: .same(proto: "failure"),
   ]
 
+  fileprivate class _StorageClass {
+    var _characteristic: CharacteristicAddress? = nil
+    var _failure: GenericFailure? = nil
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _characteristic = source._characteristic
+      _failure = source._failure
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularMessageField(value: &self._characteristic) }()
-      case 3: try { try decoder.decodeSingularMessageField(value: &self._failure) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularMessageField(value: &_storage._characteristic)
+        case 3: try decoder.decodeSingularMessageField(value: &_storage._failure)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if let v = self._characteristic {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
-    }
-    if let v = self._failure {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if let v = _storage._characteristic {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+      }
+      if let v = _storage._failure {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: WriteCharacteristicInfo, rhs: WriteCharacteristicInfo) -> Bool {
-    if lhs._characteristic != rhs._characteristic {return false}
-    if lhs._failure != rhs._failure {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._characteristic != rhs_storage._characteristic {return false}
+        if _storage._failure != rhs_storage._failure {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -1085,12 +1458,9 @@ extension NegotiateMtuRequest: SwiftProtobuf.Message, SwiftProtobuf._MessageImpl
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.deviceID) }()
-      case 2: try { try decoder.decodeSingularInt32Field(value: &self.mtuSize) }()
+      case 1: try decoder.decodeSingularStringField(value: &self.deviceID)
+      case 2: try decoder.decodeSingularInt32Field(value: &self.mtuSize)
       default: break
       }
     }
@@ -1122,37 +1492,70 @@ extension NegotiateMtuInfo: SwiftProtobuf.Message, SwiftProtobuf._MessageImpleme
     3: .same(proto: "failure"),
   ]
 
+  fileprivate class _StorageClass {
+    var _deviceID: String = String()
+    var _mtuSize: Int32 = 0
+    var _failure: GenericFailure? = nil
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _deviceID = source._deviceID
+      _mtuSize = source._mtuSize
+      _failure = source._failure
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.deviceID) }()
-      case 2: try { try decoder.decodeSingularInt32Field(value: &self.mtuSize) }()
-      case 3: try { try decoder.decodeSingularMessageField(value: &self._failure) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularStringField(value: &_storage._deviceID)
+        case 2: try decoder.decodeSingularInt32Field(value: &_storage._mtuSize)
+        case 3: try decoder.decodeSingularMessageField(value: &_storage._failure)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if !self.deviceID.isEmpty {
-      try visitor.visitSingularStringField(value: self.deviceID, fieldNumber: 1)
-    }
-    if self.mtuSize != 0 {
-      try visitor.visitSingularInt32Field(value: self.mtuSize, fieldNumber: 2)
-    }
-    if let v = self._failure {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if !_storage._deviceID.isEmpty {
+        try visitor.visitSingularStringField(value: _storage._deviceID, fieldNumber: 1)
+      }
+      if _storage._mtuSize != 0 {
+        try visitor.visitSingularInt32Field(value: _storage._mtuSize, fieldNumber: 2)
+      }
+      if let v = _storage._failure {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: NegotiateMtuInfo, rhs: NegotiateMtuInfo) -> Bool {
-    if lhs.deviceID != rhs.deviceID {return false}
-    if lhs.mtuSize != rhs.mtuSize {return false}
-    if lhs._failure != rhs._failure {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._deviceID != rhs_storage._deviceID {return false}
+        if _storage._mtuSize != rhs_storage._mtuSize {return false}
+        if _storage._failure != rhs_storage._failure {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -1166,11 +1569,8 @@ extension BleStatusInfo: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementa
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try { try decoder.decodeSingularInt32Field(value: &self.status) }()
+      case 1: try decoder.decodeSingularInt32Field(value: &self.status)
       default: break
       }
     }
@@ -1199,12 +1599,9 @@ extension ChangeConnectionPriorityRequest: SwiftProtobuf.Message, SwiftProtobuf.
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.deviceID) }()
-      case 2: try { try decoder.decodeSingularInt32Field(value: &self.priority) }()
+      case 1: try decoder.decodeSingularStringField(value: &self.deviceID)
+      case 2: try decoder.decodeSingularInt32Field(value: &self.priority)
       default: break
       }
     }
@@ -1235,32 +1632,63 @@ extension ChangeConnectionPriorityInfo: SwiftProtobuf.Message, SwiftProtobuf._Me
     2: .same(proto: "failure"),
   ]
 
+  fileprivate class _StorageClass {
+    var _deviceID: String = String()
+    var _failure: GenericFailure? = nil
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _deviceID = source._deviceID
+      _failure = source._failure
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.deviceID) }()
-      case 2: try { try decoder.decodeSingularMessageField(value: &self._failure) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularStringField(value: &_storage._deviceID)
+        case 2: try decoder.decodeSingularMessageField(value: &_storage._failure)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if !self.deviceID.isEmpty {
-      try visitor.visitSingularStringField(value: self.deviceID, fieldNumber: 1)
-    }
-    if let v = self._failure {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 2)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if !_storage._deviceID.isEmpty {
+        try visitor.visitSingularStringField(value: _storage._deviceID, fieldNumber: 1)
+      }
+      if let v = _storage._failure {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 2)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: ChangeConnectionPriorityInfo, rhs: ChangeConnectionPriorityInfo) -> Bool {
-    if lhs.deviceID != rhs.deviceID {return false}
-    if lhs._failure != rhs._failure {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._deviceID != rhs_storage._deviceID {return false}
+        if _storage._failure != rhs_storage._failure {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -1274,37 +1702,70 @@ extension CharacteristicAddress: SwiftProtobuf.Message, SwiftProtobuf._MessageIm
     3: .same(proto: "characteristicUuid"),
   ]
 
+  fileprivate class _StorageClass {
+    var _deviceID: String = String()
+    var _serviceUuid: Uuid? = nil
+    var _characteristicUuid: Uuid? = nil
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _deviceID = source._deviceID
+      _serviceUuid = source._serviceUuid
+      _characteristicUuid = source._characteristicUuid
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.deviceID) }()
-      case 2: try { try decoder.decodeSingularMessageField(value: &self._serviceUuid) }()
-      case 3: try { try decoder.decodeSingularMessageField(value: &self._characteristicUuid) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularStringField(value: &_storage._deviceID)
+        case 2: try decoder.decodeSingularMessageField(value: &_storage._serviceUuid)
+        case 3: try decoder.decodeSingularMessageField(value: &_storage._characteristicUuid)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if !self.deviceID.isEmpty {
-      try visitor.visitSingularStringField(value: self.deviceID, fieldNumber: 1)
-    }
-    if let v = self._serviceUuid {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 2)
-    }
-    if let v = self._characteristicUuid {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if !_storage._deviceID.isEmpty {
+        try visitor.visitSingularStringField(value: _storage._deviceID, fieldNumber: 1)
+      }
+      if let v = _storage._serviceUuid {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 2)
+      }
+      if let v = _storage._characteristicUuid {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 3)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: CharacteristicAddress, rhs: CharacteristicAddress) -> Bool {
-    if lhs.deviceID != rhs.deviceID {return false}
-    if lhs._serviceUuid != rhs._serviceUuid {return false}
-    if lhs._characteristicUuid != rhs._characteristicUuid {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._deviceID != rhs_storage._deviceID {return false}
+        if _storage._serviceUuid != rhs_storage._serviceUuid {return false}
+        if _storage._characteristicUuid != rhs_storage._characteristicUuid {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -1317,32 +1778,63 @@ extension ServiceDataEntry: SwiftProtobuf.Message, SwiftProtobuf._MessageImpleme
     2: .same(proto: "data"),
   ]
 
+  fileprivate class _StorageClass {
+    var _serviceUuid: Uuid? = nil
+    var _data: Data = SwiftProtobuf.Internal.emptyData
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _serviceUuid = source._serviceUuid
+      _data = source._data
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularMessageField(value: &self._serviceUuid) }()
-      case 2: try { try decoder.decodeSingularBytesField(value: &self.data) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularMessageField(value: &_storage._serviceUuid)
+        case 2: try decoder.decodeSingularBytesField(value: &_storage._data)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if let v = self._serviceUuid {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
-    }
-    if !self.data.isEmpty {
-      try visitor.visitSingularBytesField(value: self.data, fieldNumber: 2)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if let v = _storage._serviceUuid {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+      }
+      if !_storage._data.isEmpty {
+        try visitor.visitSingularBytesField(value: _storage._data, fieldNumber: 2)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: ServiceDataEntry, rhs: ServiceDataEntry) -> Bool {
-    if lhs._serviceUuid != rhs._serviceUuid {return false}
-    if lhs.data != rhs.data {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._serviceUuid != rhs_storage._serviceUuid {return false}
+        if _storage._data != rhs_storage._data {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -1356,11 +1848,8 @@ extension ServicesWithCharacteristics: SwiftProtobuf.Message, SwiftProtobuf._Mes
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try { try decoder.decodeRepeatedMessageField(value: &self.items) }()
+      case 1: try decoder.decodeRepeatedMessageField(value: &self.items)
       default: break
       }
     }
@@ -1387,32 +1876,63 @@ extension ServiceWithCharacteristics: SwiftProtobuf.Message, SwiftProtobuf._Mess
     2: .same(proto: "characteristics"),
   ]
 
+  fileprivate class _StorageClass {
+    var _serviceID: Uuid? = nil
+    var _characteristics: [Uuid] = []
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _serviceID = source._serviceID
+      _characteristics = source._characteristics
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularMessageField(value: &self._serviceID) }()
-      case 2: try { try decoder.decodeRepeatedMessageField(value: &self.characteristics) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularMessageField(value: &_storage._serviceID)
+        case 2: try decoder.decodeRepeatedMessageField(value: &_storage._characteristics)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if let v = self._serviceID {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
-    }
-    if !self.characteristics.isEmpty {
-      try visitor.visitRepeatedMessageField(value: self.characteristics, fieldNumber: 2)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if let v = _storage._serviceID {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+      }
+      if !_storage._characteristics.isEmpty {
+        try visitor.visitRepeatedMessageField(value: _storage._characteristics, fieldNumber: 2)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: ServiceWithCharacteristics, rhs: ServiceWithCharacteristics) -> Bool {
-    if lhs._serviceID != rhs._serviceID {return false}
-    if lhs.characteristics != rhs.characteristics {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._serviceID != rhs_storage._serviceID {return false}
+        if _storage._characteristics != rhs_storage._characteristics {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -1426,11 +1946,8 @@ extension DiscoverServicesRequest: SwiftProtobuf.Message, SwiftProtobuf._Message
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.deviceID) }()
+      case 1: try decoder.decodeSingularStringField(value: &self.deviceID)
       default: break
       }
     }
@@ -1459,12 +1976,9 @@ extension DiscoverServicesInfo: SwiftProtobuf.Message, SwiftProtobuf._MessageImp
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try { try decoder.decodeSingularStringField(value: &self.deviceID) }()
-      case 2: try { try decoder.decodeRepeatedMessageField(value: &self.services) }()
+      case 1: try decoder.decodeSingularStringField(value: &self.deviceID)
+      case 2: try decoder.decodeRepeatedMessageField(value: &self.services)
       default: break
       }
     }
@@ -1496,37 +2010,70 @@ extension DiscoveredService: SwiftProtobuf.Message, SwiftProtobuf._MessageImplem
     3: .same(proto: "includedServices"),
   ]
 
+  fileprivate class _StorageClass {
+    var _serviceUuid: Uuid? = nil
+    var _characteristicUuids: [Uuid] = []
+    var _includedServices: [DiscoveredService] = []
+
+    static let defaultInstance = _StorageClass()
+
+    private init() {}
+
+    init(copying source: _StorageClass) {
+      _serviceUuid = source._serviceUuid
+      _characteristicUuids = source._characteristicUuids
+      _includedServices = source._includedServices
+    }
+  }
+
+  fileprivate mutating func _uniqueStorage() -> _StorageClass {
+    if !isKnownUniquelyReferenced(&_storage) {
+      _storage = _StorageClass(copying: _storage)
+    }
+    return _storage
+  }
+
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
-    while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch fieldNumber {
-      case 1: try { try decoder.decodeSingularMessageField(value: &self._serviceUuid) }()
-      case 2: try { try decoder.decodeRepeatedMessageField(value: &self.characteristicUuids) }()
-      case 3: try { try decoder.decodeRepeatedMessageField(value: &self.includedServices) }()
-      default: break
+    _ = _uniqueStorage()
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      while let fieldNumber = try decoder.nextFieldNumber() {
+        switch fieldNumber {
+        case 1: try decoder.decodeSingularMessageField(value: &_storage._serviceUuid)
+        case 2: try decoder.decodeRepeatedMessageField(value: &_storage._characteristicUuids)
+        case 3: try decoder.decodeRepeatedMessageField(value: &_storage._includedServices)
+        default: break
+        }
       }
     }
   }
 
   func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
-    if let v = self._serviceUuid {
-      try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
-    }
-    if !self.characteristicUuids.isEmpty {
-      try visitor.visitRepeatedMessageField(value: self.characteristicUuids, fieldNumber: 2)
-    }
-    if !self.includedServices.isEmpty {
-      try visitor.visitRepeatedMessageField(value: self.includedServices, fieldNumber: 3)
+    try withExtendedLifetime(_storage) { (_storage: _StorageClass) in
+      if let v = _storage._serviceUuid {
+        try visitor.visitSingularMessageField(value: v, fieldNumber: 1)
+      }
+      if !_storage._characteristicUuids.isEmpty {
+        try visitor.visitRepeatedMessageField(value: _storage._characteristicUuids, fieldNumber: 2)
+      }
+      if !_storage._includedServices.isEmpty {
+        try visitor.visitRepeatedMessageField(value: _storage._includedServices, fieldNumber: 3)
+      }
     }
     try unknownFields.traverse(visitor: &visitor)
   }
 
   static func ==(lhs: DiscoveredService, rhs: DiscoveredService) -> Bool {
-    if lhs._serviceUuid != rhs._serviceUuid {return false}
-    if lhs.characteristicUuids != rhs.characteristicUuids {return false}
-    if lhs.includedServices != rhs.includedServices {return false}
+    if lhs._storage !== rhs._storage {
+      let storagesAreEqual: Bool = withExtendedLifetime((lhs._storage, rhs._storage)) { (_args: (_StorageClass, _StorageClass)) in
+        let _storage = _args.0
+        let rhs_storage = _args.1
+        if _storage._serviceUuid != rhs_storage._serviceUuid {return false}
+        if _storage._characteristicUuids != rhs_storage._characteristicUuids {return false}
+        if _storage._includedServices != rhs_storage._includedServices {return false}
+        return true
+      }
+      if !storagesAreEqual {return false}
+    }
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -1540,11 +2087,8 @@ extension Uuid: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementationBase,
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try { try decoder.decodeSingularBytesField(value: &self.data) }()
+      case 1: try decoder.decodeSingularBytesField(value: &self.data)
       default: break
       }
     }
@@ -1573,12 +2117,9 @@ extension GenericFailure: SwiftProtobuf.Message, SwiftProtobuf._MessageImplement
 
   mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
     while let fieldNumber = try decoder.nextFieldNumber() {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
       switch fieldNumber {
-      case 1: try { try decoder.decodeSingularInt32Field(value: &self.code) }()
-      case 2: try { try decoder.decodeSingularStringField(value: &self.message) }()
+      case 1: try decoder.decodeSingularInt32Field(value: &self.code)
+      case 2: try decoder.decodeSingularStringField(value: &self.message)
       default: break
       }
     }
